+++ CHANGELOG.md
@@ -3,6 +3,23 @@
 ## [Unreleased]
 
 
+## [v310] - 2025-09-23
+
+- Updated Poetry from 2.2.0 to 2.2.1. ([#1907](https://github.com/heroku/heroku-buildpack-python/pull/1907))
+- Updated uv from 0.8.18 to 0.8.20. ([#1910](https://github.com/heroku/heroku-buildpack-python/pull/1910))
+- Fixed errors saving the build cache when installed packages contain broken symlinks. ([#1909](https://github.com/heroku/heroku-buildpack-python/pull/1909))
+- Improved metrics for failed uv archive downloads. ([#1908](https://github.com/heroku/heroku-buildpack-python/pull/1908))
+
+## [v309] - 2025-09-19
+
+- Added metrics for misspelled `.python-version` files. ([#1904](https://github.com/heroku/heroku-buildpack-python/pull/1904))
+
+## [v308] - 2025-09-19
+
+- Updated Poetry from 2.1.4 to 2.2.0. ([#1900](https://github.com/heroku/heroku-buildpack-python/pull/1900))
+- Updated uv from 0.8.15 to 0.8.18. ([#1899](https://github.com/heroku/heroku-buildpack-python/pull/1899) and [#1901](https://github.com/heroku/heroku-buildpack-python/pull/1901))
+- Improved performance of Python build cache saving. ([#1902](https://github.com/heroku/heroku-buildpack-python/pull/1902))
+
 ## [v307] - 2025-09-10
 
 - Deprecated support for falling back to installing dependencies from a `setup.py` file if no Python package manager files were found. ([#1897](https://github.com/heroku/heroku-buildpack-python/pull/1897))
@@ -1326,7 +1343,10 @@ Default Python is now latest 2.7.10. Updated pip and Distribute.
 - Setuptools updated to v16.0
 - pip updated to v7.0.1
 
-[unreleased]: https://github.com/heroku/heroku-buildpack-python/compare/v307...main
+[unreleased]: https://github.com/heroku/heroku-buildpack-python/compare/v310...main
+[v310]: https://github.com/heroku/heroku-buildpack-python/compare/v309...v310
+[v309]: https://github.com/heroku/heroku-buildpack-python/compare/v308...v309
+[v308]: https://github.com/heroku/heroku-buildpack-python/compare/v307...v308
 [v307]: https://github.com/heroku/heroku-buildpack-python/compare/v306...v307
 [v306]: https://github.com/heroku/heroku-buildpack-python/compare/v305...v306
 [v305]: https://github.com/heroku/heroku-buildpack-python/compare/v304...v305